refactor: move users endpoints to their own file
